+++ MNSDView.pde
@@ -8,13 +8,13 @@ class MNSDView extends View {
 	QuadTree_Node<Segment2D> cselect;
 	Contour2D highlight;
 	
-	MNSDView(int sx, int sy, int ds, int cx, int cy, int tw, int th, int libsize){
+	MNSDView(int sx, int sy, float ds, int cx, int cy, int tw, int th, int libsize){
 		super(sx, sy, ds, cx, cy, tw, th, ceil(libsize/2.0));
 		
 		// Initialize Render State
 			//Color Map
 		fill = createImage(int(samplesx*spacing), int(samplesy*spacing), ARGB);
-		legend = new Legend(cornerx-22,cornery+1,12,(samplesy*spacing)-2);
+		legend = new Legend(cornerx-22,cornery+1,12,int(samplesy*spacing)-2);
 		
 			//Contours
 		contours = new ArrayList<Contour2D>();
